# src/Backlog.App/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Backlog",
    "description": "Agile Software",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Quinntyne Brown",
      "email": "quinntynebrown@gmail.com"
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Sprint/{sprintId}": {
      "get": {
        "tags": [
          "Sprint"
        ],
        "operationId": "GetSprintById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SprintId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetSprintByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sprint"
        ],
        "operationId": "RemoveSprint",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SprintId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveSprintResponse"
            }
          }
        }
      }
    },
    "/api/Sprint": {
      "put": {
        "tags": [
          "Sprint"
        ],
        "operationId": "UpdateSprint",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateSprintRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateSprintResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Sprint"
        ],
        "operationId": "GetSprints",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetSprintsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sprint"
        ],
        "operationId": "CreateSprint",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateSprintRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateSprintResponse"
            }
          }
        }
      }
    },
    "/api/Sprint/story/{storyId}": {
      "get": {
        "tags": [
          "Sprint"
        ],
        "operationId": "GetSprintsByStoryId",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "StoryId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetSprintsByStoryIdResponse"
            }
          }
        }
      }
    },
    "/api/Sprint/current": {
      "get": {
        "tags": [
          "Sprint"
        ],
        "operationId": "GetCurrentSprint",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetCurrentSprintResponse"
            }
          }
        }
      }
    },
    "/api/Sprint/story": {
      "post": {
        "tags": [
          "Sprint"
        ],
        "operationId": "AddSprintStory",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddSprintStoryRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AddSprintStoryResponse"
            }
          }
        }
      }
    },
    "/api/Sprint/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "Sprint"
        ],
        "operationId": "GetSprintsPage",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "PageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "Index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetSprintsPageResponse"
            }
          }
        }
      }
    },
...
    "/api/Story/{storyId}": {
      "get": {
        "tags": [
          "Story"
        ],
        "operationId": "GetStoryById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "StoryId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetStoryByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Story"
        ],
        "operationId": "RemoveStory",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "StoryId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveStoryResponse"
            }
          }
        }
      }
    },
    "/api/Story": {
      "get": {
        "tags": [
          "Story"
        ],
        "operationId": "GetStories",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetStoriesResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Story"
        ],
        "operationId": "CreateStory",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateStoryRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateStoryResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Story"
        ],
        "operationId": "UpdateStory",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateStoryRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateStoryDescriptionResponse"
            }
          }
        }
      }
    },
...
    "/api/Story/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "Story"
        ],
        "operationId": "GetStoriesPage",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "PageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "Index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetStoriesPageResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddSprintStoryRequest": {
      "type": "object",
      "properties": {
        "sprintId": {
          "format": "uuid",
          "type": "string"
        },
        "storyId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "AddSprintStoryResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sprint": {
          "$ref": "#/definitions/SprintDto"
        }
      }
    },
...
    "CreateSprintRequest": {
      "type": "object",
      "properties": {
        "sprint": {
          "$ref": "#/definitions/SprintDto"
        }
      }
    },
    "CreateSprintResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sprint": {
          "$ref": "#/definitions/SprintDto"
        }
      }
    },
...
    "CreateStoryRequest": {
      "type": "object",
      "properties": {
        "story": {
          "$ref": "#/definitions/StoryDto"
        }
      }
    },
    "CreateStoryResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "story": {
          "$ref": "#/definitions/StoryDto"
        }
      }
    },
...
    "GetCurrentSprintResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sprint": {
          "$ref": "#/definitions/SprintDto"
        }
      }
    },
...
    "GetSprintByIdResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sprint": {
          "$ref": "#/definitions/SprintDto"
        }
      }
    },
    "GetSprintsByStoryIdResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sprints": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SprintDto"
          }
        }
      }
    },
    "GetSprintsPageResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SprintDto"
          }
        }
      }
    },
    "GetSprintsResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sprints": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SprintDto"
          }
        }
      }
    },
...
    "GetStoriesPageResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StoryDto"
          }
        }
      }
    },
    "GetStoriesResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "stories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StoryDto"
          }
        }
      }
    },
    "GetStoryByIdResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "story": {
          "$ref": "#/definitions/StoryDto"
        }
      }
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
...
    "RemoveSprintResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sprint": {
          "$ref": "#/definitions/SprintDto"
        }
      }
    },
...
    "RemoveStoryResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "story": {
          "$ref": "#/definitions/StoryDto"
        }
      }
    },
...
    "SkillRequirementDto": {
      "type": "object",
      "properties": {
        "competencyLevel": {
          "type": "string"
        },
        "technology": {
          "type": "string"
        }
      }
    },
    "SprintDto": {
      "type": "object",
      "properties": {
        "sprintId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "start": {
          "format": "date-time",
          "type": "string"
        },
        "end": {
          "format": "date-time",
          "type": "string"
        },
        "storyIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "StoryDto": {
      "type": "object",
      "properties": {
        "storyId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "acceptanceCriteria": {
          "type": "string"
        },
        "jiraUrl": {
          "type": "string"
        },
        "dependsOn": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "status": {
          "type": "string"
        },
        "effort": {
          "format": "int32",
          "type": "integer"
        },
        "skillRequirements": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SkillRequirementDto"
          }
        }
      }
    },
...
    "UpdateSprintRequest": {
      "type": "object",
      "properties": {
        "sprint": {
          "$ref": "#/definitions/SprintDto"
        }
      }
    },
    "UpdateSprintResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sprint": {
          "$ref": "#/definitions/SprintDto"
        }
      }
    },
...
    "UpdateStoryDescriptionResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "story": {
          "$ref": "#/definitions/StoryDto"
        }
      }
    },
...
    "UpdateStoryRequest": {
      "type": "object",
      "properties": {
        "story": {
          "$ref": "#/definitions/StoryDto"
        }
      }
    },
...
  }
}